Rename symbol
Function name act( ) is rather vague because there are other functions like
move_to( ) which calls apply_action( ) so it wasn't clear what exactly takes
action. I've renamed act( ) to deliberate_action( ) that is supposed to be
the action deliberation function from poverty paper, and that is fired at each
simulation step.

Note that right now deliberation is fixed, based on the time of the day
but it should include importance and urgency in equation 3.  Currently
it just describes workplace policy with fixed schedule, so that part should
be refactored to policy declaration as well.