Add walking costs
Previously walk had a fixed effect on energy, right now it is
modified to have a variable effect based on the distance needs
to be travelled. However, at this point all agents share the same
house location and workplace location, so effectively it does not
change anything at this point.

This is just implemented for extendibility. Though I've observed that
using lambda inside the array is not a good idea because I need
to call it manuall like so:

delta = [delta[0](self._get_distance(start_coord, end_coord)), delta[1], delta[2]]
in which when needs matrix grows it will require manual labour to
adjust this usage.