Introduce needs categories
This is from a separate script I've been working on to implement
action deliberation (eqn.3 and 4). Instead of defining max capacities
as separate variables now they are introduced in category dictionary
that is special to the agent itself (that means for a randomized agent
profile, these max. capacities could differ in future work)

Also introduced need "self-esteem" that is planned to be affected by
meltdowns in which having a meltdown during work lowers Autistic
agent's self esteem