Replace wealth with financial security need (!)
I've separated wealth from needs section because agent needs some
motivation to go from home to workplace when the immediate actions
do not return any wealth. So right now if agent is at home, if going to
work is an available action, then it has this needs-satisfaction benefit of
financial security. Well, still agent is motivated to only sleep, if not, rest. even if rest effects are 0. I believe this happens because there are negative entries
in NS matrix. I'll try to fix it next.